Refactor and enhance asset conversion
This commit significantly refactors and enhances the Just Dance Editor's asset conversion capabilities, focusing on simplifying method signatures, improving error handling, and better utilizing object properties for path and data management. Key changes include the introduction of the `ConversionRequest` and `JDUbiArtSong` classes to streamline conversion requests and song data handling, respectively. The `AudioConverter.cs` and `ConvertUbiArtToUnity.cs` scripts have been updated to reduce parameter clutter, enhance version handling, and improve the efficiency of the conversion process.

+++ JustDanceEditor.Converter/UbiArt/Tapes/MusicTrack.cs
@@ -41,18 +41,18 @@ public class Structure
     public float videoStartTime { get; set; }
     public float previewEntry { get; set; }
     public float previewLoopStart { get; set; }
-    public int previewLoopEnd { get; set; }
+    public float previewLoopEnd { get; set; }
     public int volume { get; set; }
     public int fadeInDuration { get; set; }
     public int fadeInType { get; set; }
-    public int fadeOutDuration { get; set; }
+    public float fadeOutDuration { get; set; }
     public int fadeOutType { get; set; }
 }
 
 public class Signature
 {
     public string __class { get; set; }
-    public int marker { get; set; }
+    public float marker { get; set; }
     public int beats { get; set; }
 }
